feat: change isEmpty() behavior
Now this method returns true when a TVShow/Movie has its name or id
set to the default values.

This logic is followed after doing more calls to the different
themovie.org APIs, where almost all the content of the TV Show/Movie
could be empty but not the id or title.

This commit adds the changes to perform the work described above.

Relates: #MC-4