feat: add ViewModel and Repository files
In order to follow the MVVM architecture, a Repository and ViewModel
files have been added.

The Repository obtains the TV Show/Movie information from the
MainController class. On the other hand, the ViewModel gets the info
from the Repository

The information obtained is:

- List of TV Show/Movie
- List of TV Show/Movie result of searching using an specific query.

This commit adds the new files described above, along with some extra
files needed for testing or requested by Lint.

Relates: #MC-2, #MC-3