feat: add UI and refactor model package
This commit contains the following changes:

* The UI now is composed by two fragment, List and Details fragment.

The first fragment presents a list of movies/tv shows to the user. They
are easily recognizable by the top label they have.

It also contains a search option, an option to select the top rated
elements ("default ones") and another to see their watchtlist. This will
be done in a future commit.

The elements are shown on the screen by using two different RecyclerViews.

To navigate between the fragments the Navigation Component is used.
When a movies/tv show is clicked the DetailsFragment is launched with
details about the element.

Besides the details, this fragment contains a toggle button to add/
remove the element to the watchtlist.

* The model package has been refactored.

Now it only contains one class for representing a movie and a tv show.
Please refer to the model.Element class for more info about this.

Relates: #MC-6, #MC-4