feat: add watchlist option, improve UI behavior
This commit contains the following:

* Database integration using Room

Now is possible to store in a "Watch list" (stored in a SQL Database)
Movies/TV Shows the user wants to see later.

To do so, a new package (db) and methods to Repository, ViewModel and
ListFragment has been added.

In order to add/remove an element to the watchlist, click on the top
right logo is shown in its poster.

* The elements shown in the UI has changed:

Now the leftRecyclerView shows a mix of Movies/TV Shows. The type
(tv or movie) of an element can be easily identified by a top-right logo.

On the other hand, in the rightRecyclerView the WathList is shown.

When a user starts a query, the label text on top of the RecyclerViews
changes to "Movies" and "TV Shows". The elements are shown on both
RecyclerViews change to the result of searchMovie and SearchTVShow.
Please see ListFragment for more info about this.

* Added responsive layouts

fragment_list and fragment_details now have a specific layout for
landscape mode that fits the element better on the screen.

* Update documentation

The README file has been updated with information about the project.
Comments have been added in the code for better understanding.

Relates: #MC-5, #MC-6, #MC-7